Made changes to support new backend & mobile app with webapp

+++ Backend/src/main/java/com/example/demo/repository/FoundItemRepository.java
@@ -4,7 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.model.FoundItem;
 
-public interface FoundItemRepository extends JpaRepository<FoundItem, Integer> {
+public interface FoundItemRepository extends JpaRepository<FoundItem, Long> {
 }
-// This interface extends CrudRepository to provide CRUD operations for
-// FoundItem entities.
